+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * **Breaking change!** Renamed `inline` to `macro`.
 
+* Added new `-O1` optimization preset; old `-O1` became `-O2`, old `-O2` became `-O3` and so on.
+
 * Added support for parameters for macros written in Millfork.
 
 * Enabled calling macros with index expression parameters.